Create README.md
This app was created as part of the FNB App Academy. I added a twist with modern superheroes and a 3D animated, Disney-inspired style. With prior coding experience and prep through tutorials and a portfolio project, I grasped HTML and CSS easily. Creativity is key here, not perfection. As I learned, connecting the dots made sense, though I still forget some tags, I know which ones to use in the right situation.